+++ packages/ui/src/plugAndPlay/connectWallet/index.ts
@@ -1,4 +1,3 @@
 export * from "./connectWallet";
-export * from "./authorization";
 export * from "./processingTransaction";
 export * from "./extensionAccounts";
