--- packages/ui/src/readyToUse/connectWallet.tsx
@@ -1,136 +1,0 @@
-"use client";
-
-import { PropsWithChildren, useState } from "react";
-import { ExtensionStatus } from "@polkadex/react-providers";
-import { ExtensionsArray } from "@polkadot-cloud/assets/extensions";
-import { ExtensionDetails } from "@polkadex/types";
-
-import { Interaction, InteractionProps, Typography } from "../components";
-
-import { ProviderCard } from "./providerCard";
-import { SelectChain, chains } from "./selectChain";
-const ExtensionsArrayWhitelist = ExtensionsArray?.filter(
-  (item) => item.id !== "metamask-polkadot-snap"
-);
-interface ConnectWalletProps extends InteractionProps {
-  onConnectProvider: (value: ExtensionDetails) => void;
-  installedExtensions: Record<string, ExtensionStatus>;
-  onBack: () => void;
-  onConnectCallback: () => void;
-  showChains?: boolean;
-  showTerms?: boolean;
-  showFooterClose?: boolean;
-}
-export const ConnectWallet = ({
-  children,
-  onBack,
-  installedExtensions,
-  onConnectProvider,
-  onConnectCallback,
-  showChains = true,
-  showTerms = true,
-  showFooterClose,
-  ...props
-}: PropsWithChildren<ConnectWalletProps>) => {
-  const [selectedChain, setSelectedChain] = useState<(typeof chains)[0]>(
-    chains[0]
-  );
-  return (
-    <Interaction
-      className="w-full md:min-w-[24rem] md:max-w-[24rem]"
-      {...props}
-    >
-      <Interaction.Title onClose={{ onClick: onBack }}>
-        Connect your wallet
-      </Interaction.Title>
-      <Interaction.Content withPadding={false}>
-        <div className="flex flex-col gap-3">
-          {showChains && (
-            <div className="flex flex-col gap-1">
-              <Typography.Text
-                appearance="secondary"
-                size="xs"
-                className="px-7"
-              >
-                Chain
-              </Typography.Text>
-              <div className="w-full px-3">
-                <SelectChain
-                  chains={chains}
-                  onChange={(e) => setSelectedChain(e)}
-                >
-                  <SelectChain.Card
-                    title={selectedChain.name}
-                    description={selectedChain.description}
-                    icon={selectedChain.icon}
-                  />
-                </SelectChain>
-              </div>
-            </div>
-          )}
-
-          <div className="flex flex-col gap-1">
-            {showChains && (
-              <Typography.Text
-                appearance="secondary"
-                size="xs"
-                className="px-7"
-              >
-                Wallets available on the Polkadot chain
-              </Typography.Text>
-            )}
-            <div className="flex flex-col px-3 max-h-[16rem] overflow-auto">
-              {ExtensionsArrayWhitelist?.sort(
-                (a, b) =>
-                  Number(!!installedExtensions[b.id]) -
-                  Number(!!installedExtensions[a.id])
-              )?.map((value) => (
-                <ProviderCard
-                  key={value.id}
-                  title={value.title}
-                  icon={value.id}
-                  action={() => {
-                    onConnectProvider(value);
-                    onConnectCallback();
-                  }}
-                  href={(value.website as string) ?? value.website[0]}
-                  installed={!!installedExtensions?.[value.id]}
-                />
-              ))}
-            </div>
-          </div>
-          {children}
-        </div>
-      </Interaction.Content>
-      <Interaction.Footer>
-        {showTerms && (
-          <Typography.Paragraph size="xs" className="text-center">
-            By using the application, you agree to our
-            <a
-              href="https://github.com/Polkadex-Substrate/Docs/blob/master/Polkadex_Terms_of_Use.pdf"
-              target="_blank"
-              rel="noreferrer noopener"
-              className="text-primary-base align-middle"
-            >
-              {" "}
-              Terms of Service{" "}
-            </a>
-            and our{" "}
-            <a
-              href="https://github.com/Polkadex-Substrate/Docs/blob/master/Polkadex_Privacy_Policy.pdf"
-              target="_blank"
-              rel="noreferrer noopener"
-              className="text-primary-base align-middle"
-            >
-              Privacy Policy
-            </a>
-            .
-          </Typography.Paragraph>
-        )}
-        {showFooterClose && (
-          <Interaction.Close onClick={onBack}>Close</Interaction.Close>
-        )}
-      </Interaction.Footer>
-    </Interaction>
-  );
-};
