+++ packages/ui/src/readyToUse/index.ts
@@ -1,8 +1,6 @@
 export * from "./accountCard";
 export * from "./accountCombobox";
 export * from "./chainCard";
-export * from "./connectWallet";
-export * from "./extensionAccounts";
 export * from "./genericHorizontalItem";
 export * from "./hoverInformation";
 export * from "./providerCard";
